Fixed dependencies. Use factories for tests.
