+++ requirements.txt
@@ -1,4 +1,6 @@
-multidict==4.4.0
+multidict==4.5.2
+pyasn1-modules==0.2.7
+pyasn1==0.4.7
 aioxmpp>=0.10
 aiohttp==3.5.4
 aiohttp_jinja2>=0.14.0
