+++ requirements_dev.txt
@@ -7,7 +7,7 @@ pyflakes==1.6.0
 tox>2.3.1
 coverage==4.0.3
 Sphinx==1.4.8
-cryptography==2.2.2
+#cryptography==2.2.2
 pyyaml>=4.2b1
 pytest==3.6.1
 pytest-runner==2.11.1
@@ -23,3 +23,4 @@ sphinx_rtd_theme==0.4.0
 testfixtures==6.3.0
 pytest-aiohttp==0.3.0
 docutils==0.12
+factory-boy==2.12.0
