+++ tests/test_behaviour.py
@@ -4,7 +4,6 @@ from threading import Event
 
 import aioxmpp
 import pytest
-from pytest import fixture
 from asynctest import CoroutineMock, MagicMock, Mock
 
 from spade.agent import Agent
@@ -12,25 +11,12 @@ from spade.behaviour import OneShotBehaviour, CyclicBehaviour, PeriodicBehaviour
     NotValidState, NotValidTransition, BehaviourNotFinishedException
 from spade.message import Message, SPADE_X_METADATA
 from spade.template import Template
-from tests.utils import make_connected_agent, run_around_tests, wait_for_behaviour_is_killed
+from .conftest import wait_for_behaviour_is_killed
+from .factories import MockedAgentFactory
 
 STATE_ONE = "STATE_ONE"
 STATE_TWO = "STATE_TWO"
 STATE_THREE = "STATE_THREE"
-
-
-@fixture
-def message():
-    return Message(
-        to="to@localhost",
-        sender="sender@localhost",
-        body="message body",
-        thread="thread-id",
-        metadata={
-            "metadata1": "value1",
-            "metadata2": "value2"
-        }
-    )
 
 
 class StateOne(State):
@@ -55,7 +41,7 @@ class StateThree(State):
         self.kill()
 
 
-@fixture
+@pytest.fixture
 def fsm():
     fsm_ = FSMBehaviour()
     state_one = StateOne()
@@ -86,7 +72,7 @@ def test_on_start_on_end():
             self.agent.on_end_flag = True
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.on_start_flag = False
     agent.on_end_flag = False
     behaviour = TestOneShotBehaviour()
@@ -113,7 +99,7 @@ def test_on_start_exception():
         async def run(self):
             pass
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.flag = False
     behaviour = TestOneShotBehaviour()
     agent.add_behaviour(behaviour)
@@ -133,7 +119,7 @@ def test_on_run_exception():
             result = 1 / 0
             self.agent.flag = True
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.flag = False
     behaviour = TestOneShotBehaviour()
     agent.add_behaviour(behaviour)
@@ -156,7 +142,7 @@ def test_on_end_exception():
             result = 1 / 0
             self.agent.flag = True
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.flag = False
     behaviour = TestOneShotBehaviour()
     agent.add_behaviour(behaviour)
@@ -175,7 +161,7 @@ def test_add_behaviour():
         async def run(self):
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = EmptyOneShotBehaviour()
     agent.add_behaviour(behaviour)
 
@@ -200,7 +186,7 @@ def test_remove_behaviour():
         async def run(self):
             pass
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = EmptyBehaviour()
     agent.add_behaviour(behaviour)
     assert agent.has_behaviour(behaviour)
@@ -215,7 +201,7 @@ def test_remove_behaviour_not_added():
         async def run(self):
             pass
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = EmptyBehaviour()
 
     with pytest.raises(ValueError):
@@ -231,7 +217,7 @@ def test_wait_for_agent_start():
         async def run(self):
             pass
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = EmptyOneShotBehaviour()
     agent.add_behaviour(behaviour)
 
@@ -292,7 +278,7 @@ def test_send_message(message):
             await self.send(message)
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -326,7 +312,7 @@ def test_send_message_without_sender():
             await self.send(msg)
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -351,7 +337,7 @@ def test_receive():
             self.agent.recv_msg = await self.receive()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     behaviour = RecvBehaviour()
     agent.add_behaviour(behaviour)
@@ -375,7 +361,7 @@ def test_receive_with_timeout():
             self.agent.recv_msg = await self.receive(5.0)
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     msg = Message(body="received body")
     template = Template(body="received body")
@@ -403,7 +389,7 @@ def test_receive_with_timeout_error():
             self.agent.recv_msg = await self.receive(0.01)
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     template = Template(body="received body")
     behaviour = RecvBehaviour()
@@ -424,7 +410,7 @@ def test_receive_with_empty_queue():
             self.agent.recv_msg = await self.receive()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     template = Template(body="received body")
     behaviour = RecvBehaviour()
@@ -446,7 +432,7 @@ def test_set_get():
             assert self.get("key") == "value"
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -473,7 +459,7 @@ def test_multiple_templates():
             self.agent.msg3 = await self.receive(timeout=2)
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
 
     template1 = Template()
     template1.set_metadata("performative", "template1")
@@ -511,7 +497,7 @@ def test_kill_behaviour():
         async def run(self):
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = TestCyclicBehaviour()
     agent.start(auto_register=False)
 
@@ -529,7 +515,7 @@ def test_exit_code_from_kill_behaviour():
         async def run(self):
             self.kill(42)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = TestCyclicBehaviour()
     future = agent.start(auto_register=False)
     future.result()
@@ -550,7 +536,7 @@ def test_set_exit_code_behaviour():
             agent.event.wait()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     behaviour = TestCyclicBehaviour()
     agent.event = Event()
     future = agent.start(auto_register=False)
@@ -577,7 +563,7 @@ def test_notfinishedexception_behaviour():
         async def run(self):
             self.agent.event.wait()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.event = Event()
     behaviour = TestBehaviour()
     future = agent.start(auto_register=False)
@@ -605,7 +591,7 @@ def test_cyclic_behaviour():
             if self.agent.cycles > 2:
                 self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.cycles = 0
     behaviour = TestCyclicBehaviour()
     future = agent.start(auto_register=False)
@@ -628,7 +614,7 @@ def test_oneshot_behaviour():
             self.agent.one_shot_behaviour_executed = True
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.one_shot_behaviour_executed = False
     behaviour = TestOneShotBehaviour()
     agent.add_behaviour(behaviour)
@@ -649,7 +635,7 @@ def test_periodic_behaviour():
             self.agent.periodic_behaviour_execution_counter += 1
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.periodic_behaviour_execution_counter = 0
     behaviour = TestPeriodicBehaviour(0.01)
     agent.add_behaviour(behaviour)
@@ -670,7 +656,7 @@ def test_periodic_behaviour_period_zero():
             self.agent.periodic_behaviour_execution_counter += 1
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.periodic_behaviour_execution_counter = 0
     behaviour = TestPeriodicBehaviour(0)
     agent.add_behaviour(behaviour)
@@ -712,7 +698,7 @@ def test_periodic_start_at():
             self.agent.delay = datetime.datetime.now()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -736,7 +722,7 @@ def test_timeout_behaviour():
             self.agent.delay = datetime.datetime.now()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -762,7 +748,7 @@ def test_timeout_behaviour_zero():
             self.agent.delay = datetime.datetime.now()
             self.kill()
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -784,7 +770,7 @@ def test_timeout_behaviour_zero():
 
 
 def test_fsm_behaviour(fsm):
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
     agent.sync1_behaviour = asyncio.Event(loop=agent.loop)
@@ -861,7 +847,7 @@ def test_fsm_bad_state():
     fsm_.add_state(STATE_TWO, state_two)
     fsm_.add_transition(STATE_ONE, STATE_TWO)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.start(auto_register=False)
 
     assert fsm_.current_state == STATE_ONE
@@ -905,7 +891,7 @@ def test_fsm_bad_transition():
     fsm_.add_transition(STATE_ONE, STATE_TWO)
     fsm_.add_transition(STATE_TWO, STATE_THREE)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -957,7 +943,7 @@ def test_fsm_fail_on_start():
     state_one = StateOne()
     fsm_.add_state(STATE_ONE, state_one, initial=True)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -981,7 +967,7 @@ def test_fsm_fail_on_run():
     state_one = StateOne()
     fsm_.add_state(STATE_ONE, state_one, initial=True)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -1008,7 +994,7 @@ def test_fsm_fail_on_end():
     state_one = StateOne()
     fsm_.add_state(STATE_ONE, state_one, initial=True)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     future = agent.start(auto_register=False)
     future.result()
 
@@ -1034,7 +1020,7 @@ def test_join():
                 self.agent.i = i
                 await asyncio.sleep(0)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.i = None
     behaviour = WaitBehav()
 
@@ -1054,7 +1040,7 @@ def test_join_with_timeout():
         async def run(self):
             await asyncio.sleep(100)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.i = None
     behaviour = WaitBehav()
 
@@ -1074,7 +1060,7 @@ def test_join_with_long_timeout():
         async def run(self):
             await asyncio.sleep(0)
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.i = None
     behaviour = WaitBehav()
 
@@ -1100,7 +1086,7 @@ def test_join_inside_behaviour():
             await behav2.join()
             self.agent.behav1 = True
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.behav1 = False
     agent.behav2 = False
 
@@ -1131,7 +1117,7 @@ def test_join_inside_behaviour_with_timeout():
                 await behav2.join(timeout=0.001)
             self.agent.behav1 = True
 
-    agent = make_connected_agent()
+    agent = MockedAgentFactory()
     agent.behav1 = False
 
     behav1 = Behav1()
